requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
pydantic
python-dotenv
langchain
langchain-anthropic
anthropic
ddgs
httpx
beautifulsoup4